Refactor: Clean up directory structure and update code
