--- slack-ai-bot/.gitignore
@@ -1,14 +1,0 @@
-# Environment variables
-.env
-
-# Python
-__pycache__/
-*.py[cod]
-*$py.class
-
-# Logs
-*.log
-
-# ngrok
-ngrok
-ngrok-v3-stable-linux-amd64.tgz
