--- slack-ai-bot/requirements.txt
@@ -1,6 +1,0 @@
-fastapi==0.104.1
-uvicorn==0.24.0
-python-dotenv==1.0.0
-slack-sdk==3.23.0
-openai>=1.3.3
-python-multipart==0.0.6
